Feature: disallow version overwrites by policy
Prior to this change, re-uploading a package with the same version
results in the old version being overwritten.

In some environments (e.g. with packages being automatically uploaded
from "green" builds of `master`), this can result in programmer error -
forgetting to change `setup.py` versions - changing the "version of
record" in the local `pyshop`.

This change add a new administrative policy option

    pyshop.upload.never_overwrite

which results in re-uploads raising a HTTP Conflict Error.

This change also introduces `pyfakefs` to test cases, which means that
they no longer rely upon real `/tmp` or have filesystem dependencies
across tests.

+++ pyshop/views/simple.py
@@ -116,6 +116,11 @@ class UploadReleaseFile(View):
 
         filepath = os.path.join(dir_, filename)
         while os.path.exists(filepath):
+            if asbool(settings.get('pyshop.upload.never_overwrite', '0')):
+                # policy: don't overwrite files that already exist in the repo
+                raise exc.HTTPConflict(
+                    "Uploading version ({}) would overwrite existing file"
+                    .format(params['version']))
             log.warning('File %s exists but new upload self.request, deleting',
                         filepath)
             os.unlink(filepath)
